Fix data types in initialization lists.

+++ filter-newsgroups/filter-newsgroups.cpp
@@ -101,7 +101,7 @@ int main( int argc, char** argv )
                     const auto raw = mview.Raw( i );
                     fwrite( raw.ptr, 1, raw.compressedSize, ddata );
 
-                    RawImportMeta metaPacket = { offset, raw.size, raw.compressedSize };
+                    RawImportMeta metaPacket = { offset, uint32_t( raw.size ), uint32_t( raw.compressedSize ) };
                     fwrite( &metaPacket, 1, sizeof( RawImportMeta ), dmeta );
                     offset += raw.compressedSize;
